% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comment_details.R
\name{get_comment_details}
\alias{get_comment_details}
\title{Title}
\usage{
get_comment_details(id, lastModifiedDate = Sys.time(), api_keys = api_keys)
}
\arguments{
\item{id}{}

\item{lastModifiedDate}{}

\item{api_keys}{}
}
\description{
Title
}
